+++ include/hicma/util/experiment_setup.h
@@ -1,13 +1,14 @@
 #ifndef hicma_util_experiment_setup_h
 #define hicma_util_experiment_setup_h
 
+#include <cstdint>
 #include <vector>
 
 
 namespace hicma
 {
 
-std::vector<double> get_sorted_random_vector(unsigned int N);
+std::vector<double> get_sorted_random_vector(int64_t N);
 
 } // namespace hicma
