+++ src/classes/low_rank.cpp
@@ -10,6 +10,7 @@
 
 #include <algorithm>
 #include <cassert>
+#include <cstdint>
 #include <memory>
 #include <tuple>
 #include <utility>
@@ -37,13 +38,13 @@ const Dense& LowRank::S() const { return _S; }
 Dense& LowRank::V() { return _V; }
 const Dense& LowRank::V() const { return _V; }
 
-LowRank::LowRank(int m, int n, int k) : dim{m, n}, rank(k) {
+LowRank::LowRank(int64_t m, int64_t n, int64_t k) : dim{m, n}, rank(k) {
   U() = Dense(dim[0], k);
   S() = Dense(k, k);
   V() = Dense(k, dim[1]);
 }
 
-LowRank::LowRank(const Dense& A, int k) : Node(A), dim{A.dim[0], A.dim[1]} {
+LowRank::LowRank(const Dense& A, int64_t k) : Node(A), dim{A.dim[0], A.dim[1]} {
   // Rank with oversampling limited by dimensions
   rank = std::min(std::min(k+5, dim[0]), dim[1]);
   std::tie(U(), S(), V()) = rsvd(A, rank);
@@ -54,11 +55,11 @@ LowRank::LowRank(const Dense& A, int k) : Node(A), dim{A.dim[0], A.dim[1]} {
 }
 void LowRank::mergeU(const LowRank& A, const LowRank& B) {
   assert(rank == A.rank + B.rank);
-  for (int i=0; i<dim[0]; i++) {
-    for (int j=0; j<A.rank; j++) {
+  for (int64_t i=0; i<dim[0]; i++) {
+    for (int64_t j=0; j<A.rank; j++) {
       U()(i,j) = A.U()(i,j);
     }
-    for (int j=0; j<B.rank; j++) {
+    for (int64_t j=0; j<B.rank; j++) {
       U()(i,j+A.rank) = B.U()(i,j);
     }
   }
@@ -66,19 +67,19 @@ void LowRank::mergeU(const LowRank& A, const LowRank& B) {
 
 void LowRank::mergeS(const LowRank& A, const LowRank& B) {
   assert(rank == A.rank + B.rank);
-  for (int i=0; i<A.rank; i++) {
-    for (int j=0; j<A.rank; j++) {
+  for (int64_t i=0; i<A.rank; i++) {
+    for (int64_t j=0; j<A.rank; j++) {
       S()(i,j) = A.S()(i,j);
     }
-    for (int j=0; j<B.rank; j++) {
+    for (int64_t j=0; j<B.rank; j++) {
       S()(i,j+A.rank) = 0;
     }
   }
-  for (int i=0; i<B.rank; i++) {
-    for (int j=0; j<A.rank; j++) {
+  for (int64_t i=0; i<B.rank; i++) {
+    for (int64_t j=0; j<A.rank; j++) {
       S()(i+A.rank,j) = 0;
     }
-    for (int j=0; j<B.rank; j++) {
+    for (int64_t j=0; j<B.rank; j++) {
       S()(i+A.rank,j+A.rank) = B.S()(i,j);
     }
   }
@@ -86,13 +87,13 @@ void LowRank::mergeS(const LowRank& A, const LowRank& B) {
 
 void LowRank::mergeV(const LowRank& A, const LowRank& B) {
   assert(rank == A.rank + B.rank);
-  for (int i=0; i<A.rank; i++) {
-    for (int j=0; j<dim[1]; j++) {
+  for (int64_t i=0; i<A.rank; i++) {
+    for (int64_t j=0; j<dim[1]; j++) {
       V()(i,j) = A.V()(i,j);
     }
   }
-  for (int i=0; i<B.rank; i++) {
-    for (int j=0; j<dim[1]; j++) {
+  for (int64_t i=0; i<B.rank; i++) {
+    for (int64_t j=0; j<dim[1]; j++) {
       V()(i+A.rank,j) = B.V()(i,j);
     }
   }
@@ -104,14 +105,14 @@ LowRank LowRank::get_part(
   assert(row_range.start+row_range.length <= dim[0]);
   assert(col_range.start+col_range.length <= dim[1]);
   LowRank A(row_range.length, col_range.length, rank);
-  for (int i=0; i<A.U().dim[0]; i++) {
-    for (int k=0; k<A.U().dim[1]; k++) {
+  for (int64_t i=0; i<A.U().dim[0]; i++) {
+    for (int64_t k=0; k<A.U().dim[1]; k++) {
       A.U()(i, k) = U()(i+row_range.start, k);
     }
   }
   A.S() = S();
-  for (int k=0; k<A.V().dim[0]; k++) {
-    for (int j=0; j<A.V().dim[1]; j++) {
+  for (int64_t k=0; k<A.V().dim[0]; k++) {
+    for (int64_t j=0; j<A.V().dim[1]; j++) {
       A.V()(k, j) = V()(k, j+col_range.start);
     }
   }
